+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # App Center SDK for .NET Change Log
 
+## Version 2.2.0 (Under development)
+
+### App Center Data
+
+* **[Feature]** Change the default time-to-live (TTL) from 1 day to infinite (never expire).
+* **[Feature]** Add a new parameter (ReadOptions) to the list documents API.
+
+## Version 2.1.1
+
+### App Center Distribute
+
+#### iOS
+
+* **[Fix]** Fix a crash (regression from version 2.1.0) when checking for in-app updates.
+
 ## Version 2.1.0
 
 ### App Center